// src/app/views/dashboard/Dashboard.enum.ts
export enum DashboardEnumSelectTitles {
  COUNTRY = 'Country',
  NUMBER_OF_RESULTS = 'Number of Results',
  START_DATE = 'Start Date: ',
}

export enum DashboardEnumCardTitles {
  VIEWS = 'Views',
  RANK = 'Rank',
}

export enum DashboardEnumError {
  ERROR = 'Error:',
  MESSAGE = 'Please try another country or date!',
}

export enum DashboardEnumDateFormat {
  FORMAT = 'yyyy/MM/dd',
}